backend/requirements.txt
fastapi==0.111.0
alembic==1.12.0
SQLAlchemy==1.4.41
pydantic==2.7.3
pytest==7.1.3
uvicorn==0.18.3
pyjwt==2.8.0
python-multipart==0.0.9
python-dotenv==0.20.0
bcrypt==4.0.1
passlib==1.7.4
koodu==0.0.3